fix: wire theme system — modules, routes, seed migration, frontend type fixes
- Register theme model/queries/routes in mod.rs files and main.rs router
- Add Path import to themes route handler
- Add 4 theme fields to user_preferences route (light/dark slug, accent, color_mode)
- Generate 36-theme seed migration (20260221000002_themes_seed.sql)
- Fix Theme type collision in theme.service.ts (DbTheme vs local Theme)
- Add PALETTE_PRESETS/ACCENT_PRESETS exports consumed by 3 components
- Fix AccentColor import paths to theme.types.ts
- Fix applyFullTheme() to use DB theme.colors and theme.personality
- Add debounced save via Subject + switchMap, retry on API fetch
- Fix sort_order type (i16), cobalt theme missing card field

+++ backend/crates/db/src/models/theme.rs
@@ -9,7 +9,7 @@ pub struct Theme {
     pub category: String,
     pub description: String,
     pub is_dark: bool,
-    pub sort_order: i32,
+    pub sort_order: i16,
     pub is_active: bool,
     pub colors: serde_json::Value,
     pub personality: serde_json::Value,
